requirements.txt
# requirements.txt
# Climate Justice AI Orchestrator Dependencies

# Core web framework
fastapi>=0.104.0,<0.105.0
uvicorn[standard]>=0.24.0,<0.25.0

# HTTP clients
aiohttp>=3.9.0,<4.0.0
httpx>=0.25.0,<0.26.0

# Data validation and serialization
pydantic>=2.5.0,<3.0.0
pandas>=2.1.0,<2.2.0

# Configuration
python-dotenv>=1.0.0,<2.0.0

# Task queue
celery>=5.3.0,<6.0.0
redis>=5.0.0,<6.0.0

# Database
sqlalchemy>=2.0.0,<2.1.0
alembic>=1.13.0,<2.0.0

# Testing
pytest>=7.4.0,<8.0.0
pytest-asyncio>=0.21.0,<0.22.0
pytest-mock>=3.12.0,<4.0.0

# Code quality
black>=23.11.0,<24.0.0
isort>=5.12.0,<6.0.0
flake8>=6.1.0,<7.0.0
mypy>=1.7.0,<2.0.0

# Logging and monitoring
structlog>=23.2.0,<24.0.0
prometheus-client>=0.19.0,<0.20.0

# Security
cryptography>=41.0.7,<42.0.0

# Optional: Vector database for memory storage
# qdrant-client==1.7.0
# chromadb==0.4.18

# Optional: LLM clients (uncomment as needed)
# openai==1.3.9
# anthropic==0.7.8
# langchain==0.0.339